Close the existing socket connection on WriteFailed and restablish the connection.
Defect PELEDGE19-4879: pe-terminal crashes sometimes when a user tries to open a terminal
in the Portal. After enabling the debug logLevel, we found that this happens when
socket write fails with broken pipe error which leads to the termination of the program.

After further investigation, I was able to reproduce this issue only after an Internet outage is restored on the gateway.
Most likely edge-proxy closes the socket connection established by pe-terminal during network outage so next time
when pe-terminal tries to write something on the socket it throws error - broken pipe.

This fixes the issue by tearing down the connection on which it receives broken pipe error and tries
to restablish contact with the cloud.